+++ tests/unit/converter_test.py
@@ -25,7 +25,7 @@ def test_convert_iana_invalid_xml(tmp_path: Path) -> None:
 
 
 @pytest.mark.parametrize(
-    ("xml_str", "error"),
+    ("xml_str", "warning_str"),
     [
         pytest.param(
             """\
@@ -41,7 +41,10 @@ def test_convert_iana_invalid_xml(tmp_path: Path) -> None:
                     </registry>
                 </registry>
             """,
-            r'^converter: error: cannot normalize hex value "0xcejunk"',
+            (
+                'converter: warning: registry "Test registry" (id=test) skipped due to conversion '
+                'error: cannot normalize hex value "0xcejunk"\n'
+            ),
             id="value",
         ),
         pytest.param(
@@ -58,20 +61,71 @@ def test_convert_iana_invalid_xml(tmp_path: Path) -> None:
                     </registry>
                 </registry>
             """,
-            r'^converter: error: cannot normalize hex value range "0xce-0xdf"',
+            (
+                'converter: warning: registry "Test registry" (id=test) skipped due to conversion '
+                'error: cannot normalize hex value range "0xce-0xdf"\n'
+            ),
             id="range",
         ),
     ],
 )
-def test_convert_iana_invalid_hex_value(xml_str: str, error: str, tmp_path: Path) -> None:
-    with pytest.raises(RecordFluxError, match=error):
-        iana.convert(
-            data=textwrap.dedent(xml_str),
-            source=STDIN,
-            always_valid=False,
-            output_dir=tmp_path,
-            reproducible=True,
-        )
+def test_convert_iana_invalid_hex_value(
+    xml_str: str,
+    warning_str: str,
+    capsys: pytest.CaptureFixture[str],
+    tmp_path: Path,
+) -> None:
+    iana.convert(
+        data=textwrap.dedent(xml_str),
+        source=STDIN,
+        always_valid=False,
+        output_dir=tmp_path,
+        reproducible=True,
+    )
+    captured = capsys.readouterr()
+    assert captured.out == warning_str
+
+
+@pytest.mark.parametrize(
+    ("xml_str", "warning_str"),
+    [
+        pytest.param(
+            """\
+                <registry xmlns="http://www.iana.org/assignments" id="test_registry">
+                    <title>Test registry</title>
+                    <registry id="test">
+                        <title>Test registry</title>
+                        <record>
+                            <value>SOME_STRING</value>
+                            <name>SOME_NAME</name>
+                            <description>SOME_DESCRIPTION</description>
+                        </record>
+                    </registry>
+                </registry>
+            """,
+            (
+                'converter: warning: registry "Test registry" (id=test) skipped due to conversion '
+                "error: invalid literal for int() with base 10: 'SOME_STRING'\n"
+            ),
+            id="value",
+        ),
+    ],
+)
+def test_convert_iana_invalid_int_value(
+    xml_str: str,
+    warning_str: str,
+    capsys: pytest.CaptureFixture[str],
+    tmp_path: Path,
+) -> None:
+    iana.convert(
+        data=textwrap.dedent(xml_str),
+        source=STDIN,
+        always_valid=False,
+        output_dir=tmp_path,
+        reproducible=True,
+    )
+    captured = capsys.readouterr()
+    assert captured.out == warning_str
 
 
 def test_convert_iana_invalid_no_registry_id(tmp_path: Path) -> None:
@@ -291,6 +345,61 @@ def test_convert_iana_invalid_no_registry_id(tmp_path: Path) -> None:
                 end Test_Registry;
             """,
             id="simple",
+        ),
+        pytest.param(
+            """
+                <registry xmlns="http://www.iana.org/assignments" id="test_registry">
+                    <title>Test registry</title>
+                    <registry id="test1">
+                        <title>Test registry 1</title>
+                        <record>
+                            <value>101</value>
+                            <name>Value1</name>
+                        </record>
+                        <record>
+                            <value>102</value>
+                            <name>Value2</name>
+                        </record>
+                    </registry>
+                    <registry id="test2">
+                        <title>Test registry 2</title>
+                        <record>
+                            <value>201</value>
+                            <name>Value1</name>
+                        </record>
+                        <record>
+                            <value>Non-numeric value ...</value>
+                            <name>Value2</name>
+                        </record>
+                    </registry>
+                    <registry id="test3">
+                        <title>Test registry 3</title>
+                        <record>
+                            <value>301</value>
+                            <name>Value1</name>
+                        </record>
+                    </registry>
+                </registry>
+            """,
+            """\
+                -- Test registry
+                package Test_Registry is
+
+                   type Test_Registry_1 is
+                      (Value1_101 => 101,
+                       Value2 => 102)
+                   with Size => 7;
+
+                   -- type Test_Registry_2
+                   -- ... Skipped due to unsupported content in the registry ...
+
+                   type Test_Registry_3 is
+                      (Value1_301 => 301)
+                   with Size => 9;
+
+                end Test_Registry;
+            """,
+            id="with skipped registry",
         ),
     ],
 )